requirements.txt
git+https://github.com/open-data/notifications-python-client#egg=notifications-python-client
pyyaml